Update index.js
Hi Jeffrey
Please take a look at the changes I made. You had a main function within the main function that was not being called, so the code was not executing. 
Also, remember that using this API has limitations, so if you reach the limit, switch to using the mockData provided in the starter code